+++ flight/PiOS.posix/inc/pios_tcp_priv.h
@@ -57,15 +57,15 @@ typedef struct {
 	pthread_mutex_t mutex;
 	
 	pios_com_callback tx_out_cb;
-	uint32_t tx_out_context;
+	uintptr_t tx_out_context;
 	pios_com_callback rx_in_cb;
-	uint32_t rx_in_context;
+	uintptr_t rx_in_context;
 	
 	t_fifo_buffer rx_fifo;
 	uint8_t rx_buffer[PIOS_TCP_RX_BUFFER_SIZE];
 	uint8_t tx_buffer[PIOS_TCP_RX_BUFFER_SIZE];
 } pios_tcp_dev;
 
-extern int32_t PIOS_TCP_Init(uint32_t *tcp_id, const struct pios_tcp_cfg *cfg);
+extern int32_t PIOS_TCP_Init(uintptr_t *tcp_id, const struct pios_tcp_cfg *cfg);
 
 #endif /* PIOS_TCP_PRIV_H */
