+++ flight/PiOS/Common/pios_rfm22b_com.c
@@ -34,12 +34,12 @@
 #include <pios_rfm22b_priv.h>
 
 /* Provide a COM driver */
-static void PIOS_RFM22B_COM_ChangeBaud(uint32_t rfm22b_id, uint32_t baud);
-static void PIOS_RFM22B_COM_RegisterRxCallback(uint32_t rfm22b_id, pios_com_callback rx_in_cb, uint32_t context);
-static void PIOS_RFM22B_COM_RegisterTxCallback(uint32_t rfm22b_id, pios_com_callback tx_out_cb, uint32_t context);
-static void PIOS_RFM22B_COM_TxStart(uint32_t rfm22b_id, uint16_t tx_bytes_avail);
-static void PIOS_RFM22B_COM_RxStart(uint32_t rfm22b_id, uint16_t rx_bytes_avail);
-static bool PIOS_RFM22B_COM_Available(uint32_t rfm22b_com_id);
+static void PIOS_RFM22B_COM_ChangeBaud(uintptr_t rfm22b_id, uint32_t baud);
+static void PIOS_RFM22B_COM_RegisterRxCallback(uintptr_t rfm22b_id, pios_com_callback rx_in_cb, uintptr_t context);
+static void PIOS_RFM22B_COM_RegisterTxCallback(uintptr_t rfm22b_id, pios_com_callback tx_out_cb, uintptr_t context);
+static void PIOS_RFM22B_COM_TxStart(uintptr_t rfm22b_id, uint16_t tx_bytes_avail);
+static void PIOS_RFM22B_COM_RxStart(uintptr_t rfm22b_id, uint16_t rx_bytes_avail);
+static bool PIOS_RFM22B_COM_Available(uintptr_t rfm22b_com_id);
 
 /* Local variables */
 const struct pios_com_driver pios_rfm22b_com_driver = {
@@ -56,7 +56,7 @@ const struct pios_com_driver pios_rfm22b_com_driver = {
  * \param[in] rfm22b_id RFM22B name (GPS, TELEM, AUX)
  * \param[in] baud Requested baud rate
  */
-static void PIOS_RFM22B_COM_ChangeBaud(uint32_t rfm22b_id, uint32_t baud)
+static void PIOS_RFM22B_COM_ChangeBaud(uintptr_t rfm22b_id, uint32_t baud)
 {
 	struct pios_rfm22b_dev *rfm22b_dev = (struct pios_rfm22b_dev *)rfm22b_id;
 	if (!PIOS_RFM22B_validate(rfm22b_dev))
@@ -85,18 +85,18 @@ static void PIOS_RFM22B_COM_ChangeBaud(uint32_t rfm22b_id, uint32_t baud)
 	PIOS_RFM22B_SetDatarate(rfm22b_id, datarate, true);
 }
 
-static void PIOS_RFM22B_COM_RxStart(uint32_t rfm22b_id, uint16_t rx_bytes_avail)
+static void PIOS_RFM22B_COM_RxStart(uintptr_t rfm22b_id, uint16_t rx_bytes_avail)
 {
 }
 
-static void PIOS_RFM22B_COM_TxStart(uint32_t rfm22b_id, uint16_t tx_bytes_avail)
+static void PIOS_RFM22B_COM_TxStart(uintptr_t rfm22b_id, uint16_t tx_bytes_avail)
 {
 	struct pios_rfm22b_dev *rfm22b_dev = (struct pios_rfm22b_dev *)rfm22b_id;
 	if (!PIOS_RFM22B_validate(rfm22b_dev))
 		return;
 }
 
-static void PIOS_RFM22B_COM_RegisterRxCallback(uint32_t rfm22b_id, pios_com_callback rx_in_cb, uint32_t context)
+static void PIOS_RFM22B_COM_RegisterRxCallback(uintptr_t rfm22b_id, pios_com_callback rx_in_cb, uintptr_t context)
 {
 	struct pios_rfm22b_dev *rfm22b_dev = (struct pios_rfm22b_dev *)rfm22b_id;
 	if (!PIOS_RFM22B_validate(rfm22b_dev))
@@ -110,7 +110,7 @@ static void PIOS_RFM22B_COM_RegisterRxCallback(uint32_t rfm22b_id, pios_com_call
 	rfm22b_dev->rx_in_cb = rx_in_cb;
 }
 
-static void PIOS_RFM22B_COM_RegisterTxCallback(uint32_t rfm22b_id, pios_com_callback tx_out_cb, uint32_t context)
+static void PIOS_RFM22B_COM_RegisterTxCallback(uintptr_t rfm22b_id, pios_com_callback tx_out_cb, uintptr_t context)
 {
 	struct pios_rfm22b_dev *rfm22b_dev = (struct pios_rfm22b_dev *)rfm22b_id;
 	if (!PIOS_RFM22B_validate(rfm22b_dev))
@@ -124,7 +124,7 @@ static void PIOS_RFM22B_COM_RegisterTxCallback(uint32_t rfm22b_id, pios_com_call
 	rfm22b_dev->tx_out_cb = tx_out_cb;
 }
 
-static bool PIOS_RFM22B_COM_Available(uint32_t rfm22b_id)
+static bool PIOS_RFM22B_COM_Available(uintptr_t rfm22b_id)
 {
 	return PIOS_RFM22B_LinkStatus(rfm22b_id);
 }
