+++ flight/PiOS/STM32F4xx/pios_overo.c
@@ -45,10 +45,10 @@
 #define PACKET_SIZE 1024
 
 /* Provide a COM driver */
-static void PIOS_OVERO_RegisterRxCallback(uint32_t overo_id, pios_com_callback rx_in_cb, uint32_t context);
-static void PIOS_OVERO_RegisterTxCallback(uint32_t overo_id, pios_com_callback tx_out_cb, uint32_t context);
-static void PIOS_OVERO_TxStart(uint32_t overo_id, uint16_t tx_bytes_avail);
-static void PIOS_OVERO_RxStart(uint32_t overo_id, uint16_t rx_bytes_avail);
+static void PIOS_OVERO_RegisterRxCallback(uintptr_t overo_id, pios_com_callback rx_in_cb, uintptr_t context);
+static void PIOS_OVERO_RegisterTxCallback(uintptr_t overo_id, pios_com_callback tx_out_cb, uintptr_t context);
+static void PIOS_OVERO_TxStart(uintptr_t overo_id, uint16_t tx_bytes_avail);
+static void PIOS_OVERO_RxStart(uintptr_t overo_id, uint16_t rx_bytes_avail);
 
 const struct pios_com_driver pios_overo_com_driver = {
 	.set_baud   = NULL,
@@ -76,9 +76,9 @@ struct pios_overo_dev {
 	uint8_t rx_buffer[2][PACKET_SIZE];
 
 	pios_com_callback rx_in_cb;
-	uint32_t rx_in_context;
+	uintptr_t rx_in_context;
 	pios_com_callback tx_out_cb;
-	uint32_t tx_out_context;
+	uintptr_t tx_out_context;
 };
 
 #if defined(PIOS_INCLUDE_FREERTOS)
@@ -139,7 +139,7 @@ static void PIOS_OVERO_WriteData(struct pios_overo_dev *overo_dev)
  * Called at the end of each DMA transaction.  Refresh the flag indicating which
  * DMA buffer to write new data from the PIOS_COM fifo into the buffer
  */
-void PIOS_OVERO_DMA_irq_handler(uint32_t overo_id)
+void PIOS_OVERO_DMA_irq_handler(uintptr_t overo_id)
 {
 	struct pios_overo_dev * overo_dev = (struct pios_overo_dev *) overo_id;
 	if(!PIOS_OVERO_validate(overo_dev))
@@ -150,7 +150,7 @@ void PIOS_OVERO_DMA_irq_handler(uint32_t overo_id)
 	overo_dev->writing_buffer = 1 - DMA_GetCurrentMemoryTarget(overo_dev->cfg->dma.tx.channel);
 	overo_dev->writing_offset = 0;
 
-/*	bool rx_need_yield;
+/*	bool rx_need_yield = false;
 	// Get data from the Rx buffer and add to the fifo
 	(void) (overo_dev->rx_in_cb)(overo_dev->rx_in_context, 
 								 &overo_dev->rx_buffer[overo_dev->writing_buffer][0], 
@@ -173,7 +173,7 @@ void PIOS_OVERO_DMA_irq_handler(uint32_t overo_id)
 /**
  * Debugging information to check how it is runnign
  */
-int32_t PIOS_OVERO_GetPacketCount(uint32_t overo_id)
+int32_t PIOS_OVERO_GetPacketCount(uintptr_t overo_id)
 {
 	struct pios_overo_dev * overo_dev = (struct pios_overo_dev *) overo_id;
 	PIOS_Assert(PIOS_OVERO_validate(overo_dev));
@@ -184,7 +184,7 @@ int32_t PIOS_OVERO_GetPacketCount(uint32_t overo_id)
 /**
  * Debugging information to check how it is runnign
  */
-int32_t PIOS_OVERO_GetWrittenBytes(uint32_t overo_id)
+int32_t PIOS_OVERO_GetWrittenBytes(uintptr_t overo_id)
 {
 	struct pios_overo_dev * overo_dev = (struct pios_overo_dev *) overo_id;
 	PIOS_Assert(PIOS_OVERO_validate(overo_dev));
@@ -195,7 +195,7 @@ int32_t PIOS_OVERO_GetWrittenBytes(uint32_t overo_id)
 /**
  * Initialise a single Overo device
  */
-int32_t PIOS_OVERO_Init(uint32_t * overo_id, const struct pios_overo_cfg * cfg)
+int32_t PIOS_OVERO_Init(uintptr_t * overo_id, const struct pios_overo_cfg * cfg)
 {
 	PIOS_DEBUG_Assert(overo_id);
 	PIOS_DEBUG_Assert(cfg);
@@ -291,7 +291,7 @@ int32_t PIOS_OVERO_Init(uint32_t * overo_id, const struct pios_overo_cfg * cfg)
 	DMA_Cmd(overo_dev->cfg->dma.tx.channel, ENABLE);
 	DMA_Cmd(overo_dev->cfg->dma.rx.channel, ENABLE);
 
-	*overo_id = (uint32_t) overo_dev;
+	*overo_id = (uintptr_t) overo_dev;
 
 	return(0);
 	
@@ -300,7 +300,7 @@ out_fail:
 }
 
 //! Enable the SPI peripheral
-int32_t PIOS_OVERO_Enable(uint32_t overo_id)
+int32_t PIOS_OVERO_Enable(uintptr_t overo_id)
 {
 	struct pios_overo_dev * overo_dev = (struct pios_overo_dev *)overo_id;
 	
@@ -321,7 +321,7 @@ int32_t PIOS_OVERO_Enable(uint32_t overo_id)
  * - wait for BSY flag to desassert
  * this needs to be done via callbacks to avoid blocking too long
  */
-int32_t PIOS_OVERO_Disable(uint32_t overo_id)
+int32_t PIOS_OVERO_Disable(uintptr_t overo_id)
 {
 	struct pios_overo_dev * overo_dev = (struct pios_overo_dev *)overo_id;
 	
@@ -334,7 +334,7 @@ int32_t PIOS_OVERO_Disable(uint32_t overo_id)
 }
 
 
-static void PIOS_OVERO_RxStart(uint32_t overo_id, uint16_t rx_bytes_avail)
+static void PIOS_OVERO_RxStart(uintptr_t overo_id, uint16_t rx_bytes_avail)
 {
 	struct pios_overo_dev * overo_dev = (struct pios_overo_dev *)overo_id;
 	
@@ -344,7 +344,7 @@ static void PIOS_OVERO_RxStart(uint32_t overo_id, uint16_t rx_bytes_avail)
 	// DMA RX enable (enable IRQ) ?
 }
 
-static void PIOS_OVERO_TxStart(uint32_t overo_id, uint16_t tx_bytes_avail)
+static void PIOS_OVERO_TxStart(uintptr_t overo_id, uint16_t tx_bytes_avail)
 {
 	struct pios_overo_dev * overo_dev = (struct pios_overo_dev *)overo_id;
 	
@@ -357,7 +357,7 @@ static void PIOS_OVERO_TxStart(uint32_t overo_id, uint16_t tx_bytes_avail)
 	//PIOS_OVERO_WriteData(overo_dev);
 }
 
-static void PIOS_OVERO_RegisterRxCallback(uint32_t overo_id, pios_com_callback rx_in_cb, uint32_t context)
+static void PIOS_OVERO_RegisterRxCallback(uintptr_t overo_id, pios_com_callback rx_in_cb, uintptr_t context)
 {
 	struct pios_overo_dev * overo_dev = (struct pios_overo_dev *)overo_id;
 	
@@ -372,7 +372,7 @@ static void PIOS_OVERO_RegisterRxCallback(uint32_t overo_id, pios_com_callback r
 	overo_dev->rx_in_cb = rx_in_cb;
 }
 
-static void PIOS_OVERO_RegisterTxCallback(uint32_t overo_id, pios_com_callback tx_out_cb, uint32_t context)
+static void PIOS_OVERO_RegisterTxCallback(uintptr_t overo_id, pios_com_callback tx_out_cb, uintptr_t context)
 {
 	struct pios_overo_dev * overo_dev = (struct pios_overo_dev *)overo_id;
 	
@@ -389,10 +389,10 @@ static void PIOS_OVERO_RegisterTxCallback(uint32_t overo_id, pios_com_callback t
 
 #else
 
-static void PIOS_OVERO_RegisterTxCallback(uint32_t overo_id, pios_com_callback tx_out_cb, uint32_t context) {};
-static void PIOS_OVERO_RegisterRxCallback(uint32_t overo_id, pios_com_callback rx_in_cb, uint32_t context) {};
-static void PIOS_OVERO_TxStart(uint32_t overo_id, uint16_t tx_bytes_avail) {};
-static void PIOS_OVERO_RxStart(uint32_t overo_id, uint16_t rx_bytes_avail) {};
+static void PIOS_OVERO_RegisterTxCallback(uintptr_t overo_id, pios_com_callback tx_out_cb, uintptr_t context) {};
+static void PIOS_OVERO_RegisterRxCallback(uintptr_t overo_id, pios_com_callback rx_in_cb, uintptr_t context) {};
+static void PIOS_OVERO_TxStart(uintptr_t overo_id, uint16_t tx_bytes_avail) {};
+static void PIOS_OVERO_RxStart(uintptr_t overo_id, uint16_t rx_bytes_avail) {};
 
 #endif /* PIOS_INCLUDE_FREERTOS */
