+++ flight/PiOS/STM32F4xx/pios_rtc.c
@@ -39,8 +39,8 @@
 #endif
 
 struct rtc_callback_entry {
-  void (*fn)(uint32_t);
-  uint32_t data;
+  void (*fn)(uintptr_t);
+  uintptr_t data;
 };
 
 #define PIOS_RTC_MAX_CALLBACKS 3
@@ -97,7 +97,7 @@ float PIOS_RTC_MsPerTick()
 }
 
 /* TODO: This needs a mutex around rtc_callbacks[] */
-bool PIOS_RTC_RegisterTickCallback(void (*fn)(uint32_t id), uint32_t data)
+bool PIOS_RTC_RegisterTickCallback(void (*fn)(uintptr_t id), uintptr_t data)
 {
 	struct rtc_callback_entry * cb;
 	if (rtc_callback_next >= PIOS_RTC_MAX_CALLBACKS) {
